requirements.txt
alabaster==0.7.8
apt-clone==0.2.1
apturl==0.5.2
Babel==2.4.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
catfish==1.4.10
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
defer==1.0.6
Django==1.11.11
docopt==0.6.2
docutils==0.14
feedparser==5.2.1
future==0.18.2
gitsome==0.7.0
httplib2==0.9.2
idna==2.6
imagesize==0.7.1
Jinja2==2.10
louis==3.5.0
lxml==4.5.1
Mako==1.0.7
MarkupSafe==1.0
menulibre==2.2.1
mugshot==0.4.2
netifaces==0.10.4
numpy==1.19.0
numpydoc==0.7.0
onboard==1.4.1
opencv-python==4.2.0.34
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.2.0
pygobject==3.26.1
PyICU==1.9.8
pyparted==3.11.1
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
python-xapp==1.8.1
pytz==2018.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
roman==2.0.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
Sphinx==1.6.7
system-service==0.3
systemd-python==234
torch==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
wxPython==4.0.1
xkit==0.0.0
xonsh==0.6.0
